jsepos/requirement.txt
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.3
channels==1.1.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==1.2.0
Django==2.2
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msgpack-python==0.5.6
mysqlclient==1.4.2.post1
openapi-codec==1.3.2
pycparser==2.19
PyHamcrest==1.9.0
pytz==2019.1
redis==2.10.6
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.2
zope.interface==4.6.0
